Try to finish up web1999
It's been long enough, I just need to finish this post and get it out
there. It turned out to be difficult to use w9bcc-web as a page
resource seemingly because Hugo wants to treat HTML files as content
rather than resources, so I stuffed it in assets/ instead where it's
guaranteed to be treated as a resource.